% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_five.R
\docType{data}
\name{big_five}
\alias{big_five}
\title{BIG 5 Persoenlichkeitstest}
\format{Ein tible mit 200 Zeilen and 17 Spalten:
\describe{
  \item{Alter}{Alter in Lebensjahren}
  \item{Geschlecht}{f = weiblich, m = maennlich}
  \item{Herkunft}{Herkunftsland}
  \item{Extraversion}{Durchschnittliche Auspraegung von Extraversion}
  \item{Neurotizismus}{Durchschnittliche Auspraegung von Neurotizismus}
  \item{Vertraeglichkeit}{Durchschnittliche Auspraegung von Vertraeglichkeit}
  \item{Gewissenhaftigkeit}{Durchschnittliche Auspraegung von Gewissenhaftigkeit}
  \item{O1-O10}{Item 1-10 zur Auspraegung von Offenheit}
}}
\source{
\url{Research}
}
\usage{
big_five
}
\description{
Der Datensatz enthaelt drei soziodemographische Variablen, die durchschnittliche Auspraegung von Extraversion,
Neurotizismus, Vertraeglichkeit und Gewissenhaftigkeit sowie 10 Items, um die Offenheit zu messen.
}
\keyword{datasets}
