% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{Testdatensatz}
\format{A data frame with 400 rows and 3 variables:
\describe{
  \item{Geschlecht}{1 = weiblich, 2 = maennlich}
  \item{Zustand}{Empfindung ueber Glueck}
  \item{Auspraegung}{Ausmaß des Zustandes}
}}
\source{
\url{Research}
}
\usage{
df
}
\description{
Datensatz aus "indo" zum Ueben der Visualisierung mit ggpubr
}
\keyword{datasets}
