% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indo.R
\docType{data}
\name{indo}
\alias{indo}
\title{Indonesischer Datensatz ueber Lebenszufriedeneheit}
\format{A data frame with 200 rows and 11 variables:
\describe{
  \item{Geschlecht}{1 = weiblich, 2 = maennlich}
  \item{Alter}{Alter in Jahren}
  \item{Familie}{Familienstand, 1 = ledig, 2 = verheiratet}
  
}}
\source{
\url{Research}
}
\usage{
indo
}
\description{
Befragung indonesischer Probanden nach Auspraegung von Sinnerfuellung
und Sinnkrise in Form von 5 Dimensionen.
}
\keyword{datasets}
