% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indonesisch.R
\docType{data}
\name{indonesisch}
\alias{indonesisch}
\title{Indonesischer Datensatz ueber Sinnerfuellung}
\format{Ein tibble mit 200 Zeilen and 8 Spalten:
\describe{
  \item{Geschlecht}{f = weiblich, m = maennlich}
  \item{Familienstand}{ledig, verheiratet}
  \item{Alter}{Alter in Lebensjahren}
  \item{Selbsttranszendenz}{Durchschnittliche Auspraegung von Selbsttranszendenz.}
  \item{Selbstverwirklichung}{Durchschnittliche Auspraegung von Selbstverwirklichung.}
  \item{Ordnung}{Durchschnittliche Auspraegung von Ordnung.}
  \item{Wir_Wohlgefuehl}{Durchschnittliche Auspraegung von Wir Wohlgefuehl.}
  \item{Sinnerfuellung}{Durchschnittliche Auspraegung von Sinnerfuellung.}

}}
\source{
\url{Research}
}
\usage{
indonesisch
}
\description{
Befragung indonesischer Probanden nach Sinnerfuellung in Form von 4 Dimensionen.
}
\keyword{datasets}
