% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated.R
\docType{data}
\name{repeated}
\alias{repeated}
\title{Datensatz zur Messwiederholung}
\format{Ein tibble mit 90 Zeilen und 5 Spalten:
\describe{
  \item{iq}{IQ-Scores}
  \item{kreativitaet}{Kreativitaetswerte}
  \item{zeitpunkt}{Messzeitpunkt: t1, t2}
  \item{gruppe}{Gruppe: A-F}
  \item{person_id}{Identifikationsnummer}

}}
\source{
\url{Research}
}
\usage{
repeated
}
\description{
IQ-Scores und Kreativitaetswerte zu zwei Zeitpunkten von 45 in Gruppen unterteilten Probanden.
}
\keyword{datasets}
