% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipp_wm_wide.R
\docType{data}
\name{tipp_wm_wide}
\alias{tipp_wm_wide}
\title{Tipp WM (breites Format)}
\format{Ein tibble mit 64 Zeilen and 7 Spalten:
\describe{
  \item{Thomas_Bayes}{}
  \item{Egon_Pearson}{}
  \item{Jerzy_Neyman}{}
  \item{Alice}{}
  \item{Bob}{}
  \item{Charlie}{}
  \item{Ergebnis}{Wahres Ergebnis}
}}
\source{
\url{Research}
}
\usage{
tipp_wm_wide
}
\description{
WM Tipps von sechs verschiedenen Tippern in breitem Format.
}
\keyword{datasets}
