% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indonesisch.R
\docType{data}
\name{indonesisch}
\alias{indonesisch}
\title{Indonesischer Datensatz ueber Lebenszufriedeneheit}
\format{Ein tibble mit 200 Zeilen and 8 Spalten:
\describe{
  \item{Geschlecht}{f = weiblich, m = maennlich}
  \item{Familienstand}{ledig, verheiratet}
  \item{Alter}{Alter in Jahren}
  \item{Selbsttranszendenz}{}
  \item{Selbstverwirklichung}{}
  \item{Ordnung}{}
  \item{Wir_Wohlgefuehl}{}
  \item{Sinnerfuellung}{}

}}
\source{
\url{Research}
}
\usage{
indonesisch
}
\description{
Befragung indonesischer Probanden nach Auspraegung von Spiritualitaet in Form von 5 Dimensionen.
}
\keyword{datasets}
