% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree_plot.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree Plot}
\usage{
scree_plot(fa)
}
\arguments{
\item{fa}{Data.frame oder tibble mit numerischen Spalten.}
}
\value{
Scree Plot
}
\description{
Scree Plot als ggplot mit Eigenvalues auf der y-Achse
}
\examples{
\dontrun{
scree_plot(dataset)
}
}
