% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipp_wm_wide.R
\docType{data}
\name{tipp_wm_wide}
\alias{tipp_wm_wide}
\title{Tipp WM (breites Format)}
\format{Ein tibble mit 64 Zeilen and 7 Spalten:
\describe{
  \item{Thomas_Bayes}{Thomas Bayes war ein englischer Mathematiker, Statistiker, Philosoph und Pfarrer. Nach ihm ist der Satz von Bayes benannt, der in der Wahrscheinlichkeitsrechnung große Bedeutung hat.}
  \item{Egon_Pearson}{Egon Pearson war ein britischer Statistiker, der Sohn von Karl Pearson und entwickelte zusammen mit Jerzey Neyman die heute am weitesten verbreitete Theorie des Hypothesentestens.}
  \item{Jerzy_Neyman}{Jerzy Neyman war ein polnisch-US-amerikanischer Mathematiker. Das Neyman-Pearson-Lemma und der Neyman-Pearson-Test sind nach ihm benannt.}
  \item{Alice}{Alice, Strebsame Studentin}
  \item{Bob}{Bob, Statistik Nerd}
  \item{Charlie}{Charlie, Ambitionierter Sportler}
  \item{Ergebnis}{Wahres Ergebnis}
}}
\source{
\url{Research}
}
\usage{
tipp_wm_wide
}
\description{
WM Tipps von sechs verschiedenen Tippern in breitem Format.
}
\keyword{datasets}
